refactor: split internal/assets.go into focused plugin files
- plugin.go: public entry points and the embedded fallback
- plugin_download.go: HTTP download, version resolution, platform URLs
- plugin_extract.go: the five extraction formats
- plugin_info.go: metadata, hashing, validation

Complexity fixes along the way: extractFromZip (gocyclo 26) now uses
findPluginInZip/extractZipEntry helpers and one unified search over
outer-or-nested entries; extractFromPkg (17) is split into expandPkg
and findPkgPlugin; extractFromRpm loses its deep nesting via an early
return. The whole-download HTTP timeout is raised from 60s to 5m (it
covers the full ~12 MB body read) with a separate 30s bound for the
tiny VERSION check.

Closes #32
